Implement using server listener instead.
